made serialization more efficient
Network protocol used to create one new object for every serializable, while it could just write into the serializable we want to.

If we want to keep the benefit of the old behavior - which is returning a
brand new Baos with the bytes in case we want to do just that, which is
not currently in use in the project, a simple xx = new Baos();
xx.putBaos(yy); will do the trick instead of xx = yy.serializeBytes().